refactor(stella-tui): split the INSPECT payload types out of envelope.rs
Adding `Inbound::Progress` put envelope.rs over the 1500-line limit, and the
guard is right that the answer is a split rather than a baseline entry. The
INSPECT cluster (RecordedCallInfo, InspectMessage, JournalEra, InspectView) is
the one part of the envelope that is a self-contained read-model rather than a
routing tag: nothing outside it refers to these types except the two Inbound
variants that carry them, so it moves whole, with the test that pins its
80-column budget.

The deck's two init narrators move to agent/init.rs beside the stdout and
stderr ones, which keeps command_deck.rs (a grandfathered god file) under its
recorded ceiling and puts every surface's rendering of an InitLine in one
place.
